Migrate setEnabled with child propagation
The IWidget#setEnabled* methods provide overloads to propagate the new
enabled state to child widgets recursively. This child propagation used
to be the default behaviour in ancient Scout versions. Accordingly
existing code has been migrated to perform this child propagation again
(to not break any functionality).

Unfortunately this does not work well with the inheritAccessibility
feature of the widgets which may exclude children from automatic
propagation.

Therefore the former migration (performing the child propagation) is
changed to only enable/disable a box without propagating the
change to the children. Still the children are affected automatically
because a parent widget is disabled. But widgets with inherit-
Accessibility are not touched and behave correctly.

The following modifications have been performed for setEnabled and
setEnabledGranted using the child propagation:

- setEnabled(true, true, true) -> setEnabled(true, true)
  Keep the parent propagation to ensure the widgets is enabled
  afterwards.
- setEnabled(false, true, true) -> setEnabled(false)
  No parent propagation necessary as this is ignored if enabled=false.
- setEnabled(true, true, false) -> setEnabled(true, true)
  Use the default value for the child propagation.
- setEnabled(x, false, true) -> setEnabled(x)
  Remove the child propagation and use the default value of the parent
  propagation.
- setEnabled(x, true, true) -> setEnabled(x, true)
  Keep the parent propagation to ensure the widgets is enabled
  afterwards (if it is being enabled).

289438